Add user gallery photo functionality

+++ jwm-photography-api.Data/Repository/Interfaces/IPhotoRepository.cs
@@ -5,5 +5,5 @@ namespace jwm_photography_api.Data.Repository.Interfaces;
 public interface IPhotoRepository
 {
     Task<List<Photo>> GetRandomPhotosAsync(int numberOfLandscape, int numberOfPortrait, int numberOfSquare);
-    Task<bool> ExistsAsync(int id);
+    Task<bool> ExistsAsync(long id);
 }
